[#208] Rename core folder to terraform
