+++ src/commands/install/index.ts
@@ -1,7 +1,7 @@
 import { Args, Flags, Command, ux } from '@oclif/core';
 
 import { requireAwsModules } from '@/generators/addons/aws/dependencies';
-import { awsModules } from '@/generators/core/types';
+import { awsModules } from '@/generators/terraform/types';
 import { postProcess } from '@/hooks/postProcess';
 
 import { GeneralOptions } from '../generate';
