+++ src/generators/addons/aws/modules/alb.ts
@@ -9,7 +9,7 @@ import {
   INFRA_BASE_MAIN_PATH,
   INFRA_BASE_OUTPUTS_PATH,
   INFRA_BASE_VARIABLES_PATH,
-} from '@/generators/core/constants';
+} from '@/generators/terraform/constants';
 import { appendToFile, copy } from '@/helpers/file';
 
 import {
