+++ src/generators/addons/aws/modules/core/region.test.ts
@@ -1,5 +1,5 @@
 import { AwsOptions } from '@/generators/addons/aws';
-import { applyTerraformCore } from '@/generators/core';
+import { applyTerraformCore } from '@/generators/terraform';
 import { remove } from '@/helpers/file';
 
 import applyTerraformAwsProvider from './provider';
