+++ src/generators/addons/aws/modules/core/vpc.ts
@@ -6,7 +6,7 @@ import { isAwsModuleAdded } from '@/generators/addons/aws/dependencies';
 import {
   INFRA_BASE_MAIN_PATH,
   INFRA_BASE_OUTPUTS_PATH,
-} from '@/generators/core/constants';
+} from '@/generators/terraform/constants';
 import { appendToFile, copy } from '@/helpers/file';
 
 const vpcOutputsContent = dedent`
