+++ src/generators/terraform/index.ts
@@ -4,7 +4,7 @@ import { copy, rename } from '@/helpers/file';
 const applyTerraformCore = async (generalOptions: GeneralOptions) => {
   const { projectName } = generalOptions;
 
-  copy('core/', '.', projectName);
+  copy('terraform/', '.', projectName);
 
   // Need to rename .gitignore to gitignore because NPN package doesn't include .gitignore
   // https://github.com/npm/npm/issues/3763
